+++ lib/cpp/services/getPath.hpp
@@ -12,8 +12,8 @@ namespace ngPlasmid
 {
     namespace JSAware
     {
-        void getPath(const ::v8::Handle<::v8::Object>&,::ngPlasmid::Angle,::ngPlasmid::Point&,long double);
-        void getPath(const ::v8::Handle<::v8::Object>&marker,::ngPlasmid::Angle angle,::ngPlasmid::Point&center,long double trackRadius)
+        void getPath(const ::v8::Handle<::v8::Object>&,::ngPlasmid::Angle,::ngPlasmid::Point&,float);
+        void getPath(const ::v8::Handle<::v8::Object>&marker,::ngPlasmid::Angle angle,::ngPlasmid::Point&center,float trackRadius)
         {
             #ifdef PROFILE_NGPLASMID
                 PROFILER_START(JSAware::getPath);
@@ -22,7 +22,7 @@ namespace ngPlasmid
             #ifdef PROFILE_NGPLASMID
                 PROFILER_START(marker#_vadjust);
             #endif
-            long double vadjust = ::Nan::Get(
+            float vadjust = ::Nan::Get(
                 marker,
                 ::ngPlasmid::JSAware::_vadjust
             ).ToLocalChecked()->NumberValue();
@@ -30,12 +30,12 @@ namespace ngPlasmid
                 PROFILER_END();
             #endif
 
-            long double radiusInner = trackRadius + vadjust;
+            float radiusInner = trackRadius + vadjust;
 
             #ifdef PROFILE_NGPLASMID
                 PROFILER_START(marker#width);
             #endif
-            long double width = ::Nan::Get(
+            float width = ::Nan::Get(
                 marker,
                 ::ngPlasmid::JSAware::width
             ).ToLocalChecked()->NumberValue();
@@ -46,7 +46,7 @@ namespace ngPlasmid
             #ifdef PROFILE_NGPLASMID
                 PROFILER_START(marker#arrowstartwidth);
             #endif
-            long double arrowStartWidth = ::Nan::Get(
+            float arrowStartWidth = ::Nan::Get(
                 marker,
                 ::ngPlasmid::JSAware::arrowstartwidth
             ).ToLocalChecked()->NumberValue();
@@ -57,7 +57,7 @@ namespace ngPlasmid
             #ifdef PROFILE_NGPLASMID
                 PROFILER_START(marker#arrowstartlength);
             #endif
-            long double arrowStartLength = ::Nan::Get(
+            float arrowStartLength = ::Nan::Get(
                 marker,
                 ::ngPlasmid::JSAware::arrowstartlength
             ).ToLocalChecked()->NumberValue();
@@ -68,7 +68,7 @@ namespace ngPlasmid
              #ifdef PROFILE_NGPLASMID
                 PROFILER_START(marker#arrowstartangle);
             #endif
-            long double arrowStartAngle = ::Nan::Get(
+            float arrowStartAngle = ::Nan::Get(
                 marker,
                 ::ngPlasmid::JSAware::arrowstartangle
             ).ToLocalChecked()->NumberValue();
@@ -79,7 +79,7 @@ namespace ngPlasmid
             #ifdef PROFILE_NGPLASMID
                 PROFILER_START(marker#arrowendwidth);
             #endif
-            long double arrowEndWidth = ::Nan::Get(
+            float arrowEndWidth = ::Nan::Get(
                 marker,
                 ::ngPlasmid::JSAware::arrowendwidth
             ).ToLocalChecked()->NumberValue();
@@ -90,7 +90,7 @@ namespace ngPlasmid
             #ifdef PROFILE_NGPLASMID
                 PROFILER_START(marker#arrowendlength);
             #endif
-            long double arrowEndLength = ::Nan::Get(
+            float arrowEndLength = ::Nan::Get(
                 marker,
                 ::ngPlasmid::JSAware::arrowendlength
             ).ToLocalChecked()->NumberValue();
@@ -101,7 +101,7 @@ namespace ngPlasmid
              #ifdef PROFILE_NGPLASMID
                 PROFILER_START(marker#arrowendangle);
             #endif
-            long double arrowEndAngle = ::Nan::Get(
+            float arrowEndAngle = ::Nan::Get(
                 marker,
                 ::ngPlasmid::JSAware::arrowendangle
             ).ToLocalChecked()->NumberValue();
