+++ lib/cpp/services/pathArc.hpp
@@ -11,18 +11,18 @@
 namespace ngPlasmid
 {
     inline const std::string pathArc(
-        long double x,
-        long double y,
-        long double radius,
-        long double startAngle,
-        long double endAngle,
-        long double width,
-        long double arrowStartWidth,
-        long double arrowStartLength,
-        long double arrowStartAngle,
-        long double arrowEndWidth,
-        long double arrowEndLength,
-        long double arrowEndAngle
+        float x,
+        float y,
+        float radius,
+        float startAngle,
+        float endAngle,
+        float width,
+        float arrowStartWidth,
+        float arrowStartLength,
+        float arrowStartAngle,
+        float arrowEndWidth,
+        float arrowEndLength,
+        float arrowEndAngle
     ) {
         #ifdef PROFILE_NGPLASMID
             PROFILER_START(pathArc);
